Refactoring:  - included kipoi_utils and kipoi_conda package  - removed functionality and tests which are now in kipoi_utils and kipoi_conda package  - re-factored kipoi to use new packages  - added monkey patching / backward compatibility such that kipoi plugins do not note changes  - changed circle ci script to install kipoi_utils and kipoi_conda from source (they are not yet  - added pinned version of concise to test requirements
