+++ .circleci/config.yml
@@ -29,6 +29,23 @@ variables:
         pip install kipoiseq
         pip install pyfaidx
         conda install --yes -c conda-forge fastparquet zarr numcodecs
+  install_kipoi: &install_kipoi_utils
+    run:
+      name: Install Kipoi-Utils
+      command: |
+        git clone git@github.com:kipoi/kipoi-utils.git kipoi_utils_pkg
+        cd kipoi_utils_pkg
+        pip install '.'
+        cd ..
+  install_kipoi: &install_kipoi_conda
+    run:
+      name: Install Kipoi-Conda
+      command: |
+        git clone git@github.com:kipoi/kipoi-conda.git kipoi_conda_pkg
+        cd kipoi_conda_pkg
+        pip install '.'
+        cd ..
+
   install_pip_deps: &install_pip_deps
     run:
       name: Install pip dependencies
@@ -76,6 +93,8 @@ jobs:
       - *update_conda
       - *install_singularity
       - *install_pip_deps
+      - *install_kipoi_utils
+      - *install_kipoi_conda
       - *install_kipoi
       - *install_kipoi_plugins
       - *kipoi_ls
@@ -92,6 +111,8 @@ jobs:
       - checkout
       - *update_conda
       - *install_pip_deps
+      - *install_kipoi_utils
+      - *install_kipoi_conda
       - *install_singularity
       - *install_kipoi
       - *install_kipoi_plugins
