--- kipoi/conda/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import absolute_import
-
-from kipoi.conda.utils import *
-from kipoi.conda.utils import _call_and_parse, _call_command, _call_conda, _call_pip
