+++ kipoi/external/__init__.py
@@ -1,0 +1,2 @@
+from kipoi_utils.external import * # backward comp
+from . related import *
