+++ kipoi/external/related/__init__.py
@@ -1,2 +1,3 @@
-"""Custom related objects
-"""
+from kipoi_utils.external.related import * # backward comp
+from . import mixins
+from . import fields
