+++ kipoi/metadata.py
@@ -4,7 +4,7 @@ All classes inherit from `collections.Mapping` which allows to use
 `kipoi.data_utils.numpy_collate` on them (e.g. they behave as a dictionary).
 """
 from collections import Mapping
-from kipoi.data_utils import numpy_collate, numpy_collate_concat
+from kipoi_utils.data_utils import numpy_collate, numpy_collate_concat
 
 
 class GenomicRanges(Mapping):
