+++ tests/test_42_kipoi_interpret.py
@@ -16,7 +16,8 @@ from tqdm import tqdm
 from kipoi import writers, readers
 from kipoi.cli.main import prepare_batch
 import os
-from kipoi.data_utils import numpy_collate
+import kipoi_utils
+from kipoi_utils.data_utils import numpy_collate
 import numpy as np
 import copy
 from kipoi_interpret.importance_scores.gradient import Gradient
@@ -51,7 +52,7 @@ def test_score():
 
     writer = writers.HDF5BatchWriter(file_path=model.source_dir + "/example_files/grads_pred.hdf5")
 
-    with kipoi.utils.cd(model.source_dir):
+    with kipoi_utils.utils.cd(model.source_dir):
         dl = Dataloader(**dataloader_arguments)
         it = dl.batch_iter(batch_size=32, num_workers=0)
         # Loop through the data, make predictions, save the output
@@ -65,7 +66,7 @@ def test_score():
 
     obj1 = readers.HDF5Reader.load(model.source_dir + "/example_files/grads_pred.hdf5")
     obj2 = readers.HDF5Reader.load(model.source_dir + "/example_files/grads.hdf5")
-    kipoi.utils.compare_numpy_dict(obj1, obj2)
+    kipoi_utils.utils.compare_numpy_dict(obj1, obj2)
 
     if os.path.exists(model.source_dir + "/example_files/grads_pred.hdf5"):
         os.unlink(model.source_dir + "/example_files/grads_pred.hdf5")
