+++ tests/test_utils_list_files.py
@@ -1,7 +1,7 @@
 """Test 
 """
 import os
-from kipoi.utils import list_files_recursively
+from kipoi_utils.utils import list_files_recursively
 
 
 def test_list_files_recursively():
